Most of #654 (app-engine sep.) is done, finally
- The Camera class in the engine is still misplaced app code, though!...

+++ src/app/model/Cfg.hpp
@@ -3,7 +3,7 @@
 
 namespace OON::Model {
 
-	using BasicNumberType = double;
+	using BasicNumberType = float;
 
 }
